scripts/eval-generator/requirements.txt
PyMuPDF
Pillow
openai
python-dotenv
